DTO package moved to model
