+++ src/main/java/org/example/pmas/model/dto/ProjectDTO.java
@@ -1,4 +1,4 @@
-package org.example.pmas.dto;
+package org.example.pmas.model.dto;
 
 import java.util.Date;
