+++ src/main/java/org/example/pmas/repository/Interfaces/IProjectRepository.java
@@ -1,6 +1,6 @@
 package org.example.pmas.repository.Interfaces;
 
-import org.example.pmas.dto.ProjectDTO;
+import org.example.pmas.model.dto.ProjectDTO;
 import org.example.pmas.model.Project;
 import org.springframework.stereotype.Repository;
