+++ src/main/java/org/example/pmas/repository/ProjectRepository.java
@@ -1,6 +1,6 @@
 package org.example.pmas.repository;
 
-import org.example.pmas.dto.ProjectDTO;
+import org.example.pmas.model.dto.ProjectDTO;
 import org.example.pmas.exception.DatabaseException;
 import org.example.pmas.model.Project;
 import org.example.pmas.model.rowMapper.ProjectDTORowMapper;
